refactor(screens): move fragments to package
