+++ android/app/src/main/res/layout/first_fragment.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".FirstFragment">
+  tools:context=".screens.FirstFragment">
 
   <TextView
     android:id="@+id/first_textview"
