+++ android/app/src/main/res/layout/second_fragment.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".SecondFragment">
+  tools:context=".screens.SecondFragment">
 
   <TextView
     android:id="@+id/second_textview"
